+++ src/ui/svelte/fancy-toolbar/state/system.svelte.ts
@@ -3,7 +3,13 @@ import { currentMonitorId, monitors, mousePos, virtualDesktop } from "./getters.
 
 export { virtualDesktop };
 
-const _currentMonitor = $derived(monitors.value.find((m) => m.id === currentMonitorId)!);
+const _currentMonitor = $derived.by(() => {
+  const monitor = monitors.value.find((m) => m.id === currentMonitorId);
+  if (!monitor) {
+    throw new Error("Current monitor not found");
+  }
+  return monitor;
+});
 
 const _mouseAtEdge = $derived.by((): FancyToolbarSide | null => {
   const box = _currentMonitor.rect;